jiraAPIKit/requirement.txt
pysqlite3-binary
numpy
pandas
dotmap
scikit-learn
boto3
seaborn
matplotlib
plotly
Werkzeug
dash
dash_extensions
dash-bootstrap-components
Flask-Cors
load_dotenv
langchain
langchain-openai
python-decouple
langchainhub
langchain_groq
langgraph
ipython
pyowm
langchain_huggingface
atlassian-python-api
streamlit-mermaid
streamlit
langchain_community
chromadb
langchain_chroma